Add files via upload
This exercise employs ANOVA in a linear model with the objective to analyse the effect of baking powder amount (indepndent variable) in a biscuit dough (the experimental unit) upon the height it rises (Response). [Hypothetical data, for the pupose of learning]. 
As can be seen from the table in the csv file, there are four levels of baking powder. The four replicates of the four levels were made in a randomized order.